+++ datacontract/templates/style/tailwind.config.js
@@ -1,5 +1,15 @@
 module.exports = {
-  content: ["../datacontract.html", "../index.html", "../partials/model_field.html", "../partials/definition.html", "../partials/datacontract_information.html", "../partials/datacontract_servicelevels.html", "../partials/datacontract_terms.html", "../partials/example.html"],
+  content: [
+    "../datacontract.html",
+    "../index.html",
+    "../partials/model_field.html",
+    "../partials/server.html",
+    "../partials/definition.html",
+    "../partials/datacontract_information.html",
+    "../partials/datacontract_servicelevels.html",
+    "../partials/datacontract_terms.html",
+    "../partials/example.html",
+  ],
   theme: { },
   plugins: [],
 }