+++ tests/test_import_bigquery.py
@@ -1,11 +1,11 @@
 import logging
-from unittest.mock import patch
 
 import yaml
 from typer.testing import CliRunner
 
 from datacontract.cli import app
 from datacontract.data_contract import DataContract
+from datacontract.imports.bigquery_importer import import_bigquery_from_json
 
 logging.basicConfig(level=logging.DEBUG, force=True)
 
@@ -35,51 +35,16 @@ def test_import_bigquery_schema():
     assert DataContract(data_contract_str=expected).lint(enabled_linters="none").has_passed()
 
 
-@patch("google.cloud.bigquery.Client.get_table")
-def test_import_from_api(mock_client):
-    # Set up mocks
-    # mock_table = Mock()
-    # mock_table.to_api_repr.return_value = {
-    #     'kind': 'bigquery#table',
-    #     'etag': 'K9aCQ39hav0KNzG9cq3p7g==',
-    #     'id': 'bigquery-test-423213:dataset_test.table_test',
-    #     'selfLink': 'https://bigquery.googleapis.com/bigquery/v2/projects/bigquery-test-423213/datasets/dataset_test/tables/table_test',
-    #     'tableReference': {'projectId': 'bigquery-test-423213',
-    #                         'datasetId': 'dataset_test',
-    #                         'tableId': 'table_test'},
-    #     'description': 'Test table description',
-    #     'labels': {'label_1': 'value_1'},
-    #     'schema': {'fields': [{'name': 'field_one',
-    #                             'type': 'STRING',
-    #                             'mode': 'REQUIRED',
-    #                             'maxLength': '25'},
-    #                         {'name': 'field_two',
-    #                             'type': 'INTEGER',
-    #                             'mode': 'REQUIRED'},
-    #                         {'name': 'field_three',
-    #                             'type': 'RANGE',
-    #                             'mode': 'NULLABLE',
-    #                             'rangeElementType': {'type': 'DATETIME'}}]},
-    #     'numBytes': '0',
-    #     'numLongTermBytes': '0',
-    #     'numRows': '0',
-    #     'creationTime': '1715626217137',
-    #     'expirationTime': '1720810217137',
-    #     'lastModifiedTime': '1715626262846',
-    #     'type': 'TABLE',
-    #     'location': 'europe-west6',
-    #     'numTotalLogicalBytes': '0',
-    #     'numActiveLogicalBytes': '0',
-    #     'numLongTermLogicalBytes': '0'}
+def test_import_multiple_bigquery_schemas_with_different_types():
+    result = DataContract().import_from_source("bigquery", "fixtures/bigquery/import/multi_import_table.json")
 
-    # mock_client.response_value = mock_table
+    result = import_bigquery_from_json(result, "fixtures/bigquery/import/multi_import_external_table.json")
+    result = import_bigquery_from_json(result, "fixtures/bigquery/import/multi_import_snapshot.json")
+    result = import_bigquery_from_json(result, "fixtures/bigquery/import/multi_import_view.json")
+    result = import_bigquery_from_json(result, "fixtures/bigquery/import/multi_import_materialized_view.json")
 
-    # Call the API Import
-    # result = DataContract().import_from_source(format="bigquery", source=None, tables=["Test_One"], bt_project_id=
-    #    "project_id", bt_dataset_id="dataset_id")
-    # print("Result:\n", result)
-    # TODO: This really should have a proper test, but I've not been able to set the mocks up
-    # correctly – maybe there's some help to be had?
-    # Anyway, the serialized dict above is a real response as captured from the Bigquery API and should
-    # be sufficient to check the behavior of this way of importing things
-    assert True is True
+    print("Result:\n", result.to_yaml())
+    with open("fixtures/bigquery/import/datacontract_multi_import.yaml") as file:
+        expected = file.read()
+    assert yaml.safe_load(result.to_yaml()) == yaml.safe_load(expected)
+    assert DataContract(data_contract_str=expected).lint(enabled_linters="none").has_passed()