+++ tests/test_import_odcs.py
@@ -10,6 +10,7 @@ from datacontract.data_contract import DataContract
 
 logging.basicConfig(level=logging.DEBUG, force=True)
 
+
 def test_cli():
     runner = CliRunner()
     result = runner.invoke(
@@ -24,11 +25,13 @@ def test_cli():
     )
     assert result.exit_code == 0
 
+
 def test_import_full_odcs():
     result = DataContract().import_from_source("odcs", "./fixtures/odcs/odcs-full-example.yaml")
     expected_datacontract = read_file("fixtures/odcs/full-example.datacontract.yml")
     assert yaml.safe_load(result.to_yaml()) == yaml.safe_load(expected_datacontract)
     assert DataContract(data_contract_str=expected_datacontract).lint(enabled_linters="none").has_passed()
+
 
 def test_import_complex_odcs():
     result = DataContract().import_from_source("odcs", "./fixtures/odcs/odcs-postgresql-adventureworks-contract.yaml")